react-components: Remove 'connectActive' method
Instead of using this method, one should just call `connect` directly on the connector which will then notify `WithWalletConnector` via its role as the connector's delegate. The application calls `setActiveConnection` to set it as active.

This means that `activeConnectorError` no longer holds errors for connections and no longer maintains the `isConnecting` indicator as it's now the application's job to track this state. It's possible that the `activeConnection` field will be moved away from this component next.

As part of this change, the `onConnected` event was expanded to include the initially connected account.