CI: Add release workflow
The workflow is triggered manually with the following parameters:

- Use workflow from (always present): The branch/tag that the workflow runs from.
- Tag: The tag to build from. Defaults to Use workflow from which must then be a tag.
- Release notes: The textual body of the release. It turns out that Actions doesn't support multiline inputs, so only a single line can be provided. The intent is to find a way to automatically pick the relevant snippet from the correct changelog, but this isn't implemented yet.

The workflow checks the tag is of the form '<project>/<version>-<build-version>' where '<project>' matches a directory in './packages' and '<version>' matches the version field in that project's 'package.json' file.

It then checks out the code, builds '<project>', and uploads the zipped contents of './dist' as a release with name '<project>: <version>-<build-version>'. As mentioned above, for now, the release notes can only consist of the single input line.