Add support for signing both string and binary messages
Updated the 'sign-message' sample dApp to support both variants.

The mobile wallets don't yet support signing binary messages, so the WalletConnect implementation throws an exception for that case. It's very much not desired to throw exceptions at use time for unimplemented features as compared to enabling the dApp to detect available features for a given connection. But in this case, the necessary "warts" in the interface is not deemed worth it: It's more like the MW isn't lacking the feature, only part of it.