contractupdate: Use gRPC client
The JSON-RPC client is deprecated and replaced with the gRPC client. When merging that change, only one of the RPC usages in the sample dApp were migrated. This migrates the other use (to use the same client). As the new client is connection-independent, the query screen now appears immediately after the application has loaded.

A significant change is that the 'getModuleSource' method of gRPC client returns the module source without the module version and length (and actually another byte before that) prefix.