wallet-connectors: WalletConnect: Use "lower" camelCase format for transaction type
This apparently is the standard style used for these strings. We have to bypass TypeScript type checking in order to convert the `AccountTransactionType` enum to `TransactionKindString` (which is assumed to use the canonical naming) as it cannot infer that the two enums use the same keys (I suppose we should add a test in @concordium/common-sdk to ensure that it stays that way).

This doesn't change anything in the Android wallet as it doesn't read the field but just assumes that it's an `update` transaction. The iOS implementation reads the field and will support both formats for backwards compatibility with projects using old versions of @concordium/wallet-connectors. The Android implementation must do the same when it starts supporting other transaction types.