useConnect: Make connect function undefined until connector is ready
Let `connection` function be `undefined` if `connector` is. This prevents the function from being called before `WalletConnectionProps.activeConnector` is ready which would fail anyway. The expectation before was that the button/function invoking `connect` would check this itself, but making it explicit in the type seems less error-prone.

I do not really consider this a breaking change because the type is being strictly expanded (to include `undefined`) and doesn't break any existing correct code. If one did pass the `connect` function to other functions (which would be weird), the type signatures of those functions might have to be updated though.

An alternative solution would be to return an additional boolean `isReadyToConnect`, but that would not allow the type system to prevent the call like it does with this one.